split out lexer functions
splits out lexer functions into src/parser/lexer.cc/h. Separates the
sqltoast::parse() implementation into src/parser/parse.cc.
